fixed VS2005 compiler warning
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@884 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ tests/cv/src/acamshift.cpp
@@ -495,7 +495,7 @@ int CV_MeanShiftTest::validate_test_results( int /*test_case_idx*/ )
         goto _exit_;
     }
 
-    if( fabs(comp.area - area0) > fabs(area - area0) + area0*0.05 )
+    if( fabs((double)(comp.area - area0)) > fabs((double)(area - area0)) + area0*0.05 )
     {
         ts->printf( CvTS::LOG,
             "Incorrect CvConnectedComp area (=%.1f, should be %d)\n", comp.area, area0 );
